+++ spec/dummy/app/models/dummy_model.rb
@@ -1,7 +1,4 @@
 class DummyModel < ApplicationRecord
-  include Agilibox::Search
-  include Agilibox::TimestampHelpers
-
   belongs_to :asso, polymorphic: true, required: false
 
   polymorphic_id_for :asso